% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_group.R
\name{compare_group}
\alias{compare_group}
\title{Compare variables in two groups}
\usage{
compare_group(GCMlasso_obj = GCMlasso_obj, grp1, grp2, var,
  credible_level = 0.95)
}
\arguments{
\item{GCMlasso_obj}{\code{GCMlasso} object.}

\item{grp1}{levels in the cluster variable that belong to group1.}

\item{grp2}{levels in the cluster variable that belong to group2.}

\item{var}{variables to be compared.}

\item{credible_level}{level of the credible interval.}
}
\value{
Posterior means and credible intervals of the differences
of variables in two groups.
}
\description{
Compute the differences of the random effects for variables from two groups.
}
\examples{
compare_group(GCMlasso_obj=GCMlasso_obj,grp1=1:2,grp2=3:4,
 var=1:15,credible_level=0.95)
}
