% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCMlasso.R
\name{GCMlasso}
\alias{GCMlasso}
\title{Gaussian graphical model for ordered variables}
\usage{
GCMlasso(data, var_ord, var_group, nsamp = 1000, odens = 1,
  nwarm = 500, seed = 1, s = 0.01, t = 0.01, verb = TRUE)
}
\arguments{
\item{data}{an N by p data frame,}

\item{var_ord}{indices of ordinal variables.}

\item{var_group}{index of the variable that defines clusters.}

\item{nsamp}{number of MCMC iterations.}

\item{odens}{number of iterations between saved samples.}

\item{nwarm}{number of MCMC iterations as burn-in.}

\item{seed}{a random integer.}

\item{s}{hyperparameter of \eqn{lambda}, degree of freedom in adaptive graphical lasso prior.}

\item{t}{hyperparameter of \eqn{lambda}, shrinkage in adaptive graphical lasso prior.}

\item{verb}{print progress of MCMC, logical TRUE or FALSE.}
}
\value{
An object with S3 class "\code{GCMlasso}" is returned.

\item{data_ordered}{the same as the input data but ordered by the
clustering variable.}

\item{Gamma.st}{saved variance covariance matrices for latent variables.}

\item{Omega.st}{saved precision matrices for latent variables.}

\item{psi.st}{saved variance covariance matrices for random effects.}

\item{z.st}{saved latent variables.}

\item{b.st}{saved random effects.}
}
\description{
Perform Bayesian Gaussian graphical model for ordered variables
with clustering structure. The conditional independence between variables
are measured on the latent scale via the extended rank likelihood method.
Shrinkage effects are applied on the precision matrix to handle
multicollinearity. Clustering effects are modelled through the random effects.
Missing data are allowed.
}
\details{
\code{GCMlasso} function fits the Bayesian Gaussian copula model with graphical lasso prior
for variables with ordering (continuous, ordinal and binary) in multilevel data sets.
Adaptive graphical lasso prior is put on the precion matrix of the latent variables
conditional on the random effects, where the latent variables are implied by the
extended rank likelihood method.

\code{var_group} is the index of the varible that defines the clusters, and
should be placed at the last column in the \code{data}. The coding for the clustering variable
is from 1 to the total number of clusters. The binary variables
in \code{var_ord} should be coded as 0 (control) or 1(case) in the case-control studies.

Missing data are allowed for ordered variables and should be denoted as NA.
}
\examples{
GCMlasso_obj<-GCMlasso(data=Framingham,var_ord=1:15,var_group=16,
  nsamp=1000,odens=1,nwarm=500,seed=1,s=1e-2,t=1e-2,verb=TRUE)
}
\references{
\insertRef{hoff2007extending}{GCMlasso}

\insertRef{wang2012bayesian}{GCMlasso}

\insertRef{wang2017copula}{GCMlasso}
}
\author{
Jiali Wang (\email{jiali.wang@data61.csiro.au})
}
