% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot graphical model}
\usage{
plot_graph(GCMlasso_obj = CMlasso_obj, var, edge_perc, seed = 1)
}
\arguments{
\item{GCMlasso_obj}{\code{GCMlasso} object.}

\item{var}{variables in the graph.}

\item{edge_perc}{cut-off quantile value of edge inclusion.}

\item{seed}{a random integer.}
}
\description{
Plot the conditional dependence between variables on the latent scale.
The assocations between variables are computed from the precision matrices
from the MCMC, and the percentage of edges to be included in the graph
is determined by the \code{edge_perc} value.
}
\examples{
plot_graph(GCMlasso_obj,var=1:15,edge_perc=0.65)

}
