% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict positive outcomes}
\usage{
predict(GCMlasso_obj, var_response, var_group, rep_sample = 100,
  seed = 1)
}
\arguments{
\item{GCMlasso_obj}{\code{GCMlasso} object.}

\item{var_response}{index of binary response variable.}

\item{var_group}{index of the variable that defines clusters.}

\item{rep_sample}{number of samples in the posterior predictive sampling.}

\item{seed}{a random integer.}
}
\value{
Mean values of posterior predictive probabilities of positive cases
for observations with known outcomes and new observations.
}
\description{
Making predictions for positive outcomes for binary response.
}
\details{
Include new observations in the data set by treating the predicting variables
as NA, and run the MCMC using the function \code{GCMlasso}. Code positive
cases as 1 and controls as 0. The maximum latent variable corresponding to the
positive cases is used as threshold to predict new observations.
}
\examples{
predict_val<-predict(GCMlasso_obj,var_response=15,var_group=16)
}
