% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{reg_coef}
\alias{reg_coef}
\title{Compute regression coefficients}
\usage{
reg_coef(GCMlasso_obj, var_pred, var_response, credible_level = 0.95)
}
\arguments{
\item{GCMlasso_obj}{\code{GCMlasso} object.}

\item{var_pred}{indices of variables as predictors.}

\item{var_response}{index of variables as response.}

\item{credible_level}{level of credible interval.}
}
\value{
Posterior means and credible intervals of the coefficients.
}
\description{
Compute the regression coefficients of the predictors on a response on the
latent variables scale.
}
\examples{
reg_coef(GCMlasso_obj,var_pred=1:14,var_response=15)

}
