% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinvgauss.R
\name{rinvgauss}
\alias{rinvgauss}
\title{Sample from inverse Gaussian distribution}
\usage{
rinvgauss(n, mean = 1, shape = NULL, dispersion = 1)
}
\arguments{
\item{n}{sample size.}

\item{mean}{vector of positive means.}

\item{shape}{vector of (positive) shape parameters.}

\item{dispersion}{vector of (positive) dispersion parameters.}
}
\value{
A vector of random samples from inverse Gaussian distribution
}
\description{
Draw samples from inverse Gaussian distribution
}
\examples{
rinvgauss(100,mean=1,shape=2)
}
